[Feature] Added form validation

+++ client/userAuth/userAuthController.js
@@ -1,28 +1,62 @@
 angular.module('menuApp')
 
-.controller('userAuthController', ['$window', '$scope', 'Auth', '$location', 'Preferences', '$state', function($window, $scope, Auth, $location, Preferences, $state) {
+.controller('userAuthController', ['$window', '$scope', 'Auth', 'ValidationFactory', '$location', '$state',function($window, $scope, Auth, ValidationFactory, $location, $state) {
 
   $scope.signUp = function() {
+    var validate = ValidationFactory.validatePasswordAndEmail($scope.user.email, $scope.user.password, $scope.user.confirmpassword);
+    console.log($scope.user.password === $scope.user.confirmpassword);
+    console.log(validate);
+    if(validate.passwordFormFlag === false && validate.emailFormFlag === false) {
+      swal({
+        title: '',
+        text: 'Password and/or email are not valid! Please doublecheck and resubmit.',
+        type: 'error',
+        confirmationButtonText: 'OK'
+      });
+    } else if(validate.passwordFormFlag === false) {
+      swal({
+        title: '',
+        text: 'Passwords do not match! Please re-enter your passwords and make sure they match.',
+        type: 'error',
+        confirmationButtonText: 'OK'
+      });
+    } else if(validate.emailFormFlag === false) {
+      swal({
+        title: '',
+        text: 'Invalid email! Please enter a valid email address.',
+        type: 'error',
+        confirmationButtonText: 'OK'
+      });
+    } else {
     Auth.signup($scope.user)
-    .then(function(token) {
-      $window.localStorage.setItem('authentication', token);
-      console.log('LOOK AT YOU, SIGNING UP AND WHATNOT. WHAT A BIG BOY.');
-      $state.go('profileView');
-      Preferences.showTabDialog();
-    })
-    .catch(function(error) {
-      console.log(error);
-    });
+      .then(function(token) {
+        $window.localStorage.setItem('authentication', token);
+        $location.path('/user');
+      })
+      .catch(function(error) {
+        swal({
+          title: 'Username already exists!',
+          text: 'Please choose a different username or sign into your account.',
+          type: 'error',
+          confirmationButtonText: 'OK'
+        });
+      });
+    }
   };
 
   $scope.validate = function() {
     Auth.signin($scope.user)
     .then(function(token) {
       if(token === undefined) {
-        console.log('INCORRECT LOGIN');
+        swal({
+          title: '',
+          text: 'Username/password do not exist! Double check your credentials or create a new account.',
+          type: 'error',
+          confirmationButtonText: 'OK'
+        });
+        $location.path('/signIn');
       } else {
         $window.localStorage.setItem('authentication', token);
-        console.log('SUCCESS!!!!');
         $state.go('restaurantList');
       }
     });
